refactor(test): split workflow test files into focused modules
Split oversized workflow test files by responsibility to improve
maintainability while preserving coverage:

- test_engine.py (648L) → test_engine_execution.py + test_engine_submission.py
  + _engine_helpers.py (FakeExecutable, install_fake_builder, etc.)
- test_executable.py (443L) → test_executable_agent.py + test_executable_function.py
  + test_executable_builder.py + _executable_helpers.py + conftest.py (team_deps fixture)

Helpers prefixed with underscore so pytest's test_*.py discovery skips them.

Created by 🤖Claude Opus 4.7
